Added transaction code
this change creates a transaction at the begning of the Aouth flow.  This allows us to create a redirect with only one parameter (transactionId) that is then used to re-read the parameters sent to the connectService call.  It will allow us to expand the parameters without changing the redirection page.