polar improvements to signature and fit file
this change corrects the calculation of the polar webhook signature so that it finds a match and can validate the message.  if no match is found it still defaults to project rather than fail.  TODO is to take out this and fail if no match

The FIT files that are downloaded were not readable.  This change fixes the request headers and uses rawBody to save the information rather than using body.  In my test this allows the file to become readable when downloaded.