Improved the retry functionality and the webhook suppression
This change tests that the webhook suppression field works as expected.

It also improves the retry functionality so that retries are now working.  However, any wait for longer than 60s times out the function so we will need to revisit this with a second change to trigger a scheduled function.