Tests of new onCreate Activity function
This change introduces the onCreate Activity function for sending webhook messages to developers.  The change also includes changes to the webhook tests to ensure they check the function more correctly.  still havfe a little issue where the test can report an error if the cloud function is triggered before the local function completes.  But it runs successfully in the wild