update Wahoo Keys to production
This change updates the wahoo keys and secret to the production keys. To release we need to change the url's in the wahoo application on the wahoo management site at the same time as we release them these keys to the live environment.  So this needs a bit of downtime/quite time to do it and needs to be co-ordinated.