add client ID to userDoc to make it easier to update refreshed codes
when refreshing access codes we need to update every userDoc that for the same provider user id who has been authorised with a specific client_id.  This requires the cleint_id to be on the userDoc record.  It will massivly simplify updating access codes when they are refreshed without updating the wrong records