update config for Wahoo
This change introduces a new configuration property "providerConfigs".  This gives an array of provider secrets with their matching client ID and webhookIDs.

This provides a simpler wat to lookup the secrets if we have one of the matching other ids.

For wahoo this is now the way we manage secrets. changes to the other providers will be processed later.