update secret lookups
Secrets should be gotten by looking up the secrets relevant to the <provider>_client_id on the userDoc.  This change makes this consistent.  The tag on the developer doc is used for connecting services but once the service is connects the userDoc has the client ID which is used to get the secrets that were used in the first place.